flask_signoz_demo/requirements.txt
Flask==2.3.2
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-exporter-otlp==1.18.0
python-dotenv==1.0.0
requests==2.31.0
